api/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
python-dotenv>=1.0
motor>=3.4
pydantic>=2.7
pydantic-core>=2.18
typing-extensions>=4.12
